Stream-Three/django-todo/part_5/django_todo/requirements.txt
appdirs==1.4.3
Django==1.11
django-cors-headers==2.0.2
djangorestframework==3.6.2
djangorestframework-jwt==1.8.0
drfdocs==0.0.11
packaging==16.8
PyJWT==1.4.2
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
gunicorn
whitenoise
coreapi
pygments
markdown
coreapi-cli
dj-database-url==0.2.1
MySQL-python==1.2.5